airflow-splunk-integration/requirements.txt
apache-airflow
requests
psycopg2-binary